Extract types and interface to separate files
