+++ src/types/ErrorResult.ts
@@ -1,0 +1,4 @@
+export default interface ErrorResult {
+  name?: string;
+  price?: string;
+}
