+++ src/types/ItemFormProps.ts
@@ -1,0 +1,3 @@
+export default interface ItemFormProps {
+  toggleForm: () => void;
+}
