+++ src/types/ItemListProps.ts
@@ -1,0 +1,3 @@
+export default interface ItemListProps {
+  whetherBought: boolean;
+}
