+++ src/types/ItemProps.ts
@@ -1,0 +1,6 @@
+// Project files
+import ShoppingItem from "../types/ShoppingItem";
+
+export default interface ItemProps {
+  item: ShoppingItem;
+}
