+++ src/types/SortOptionContent.ts
@@ -1,0 +1,4 @@
+export default interface SortOptionContent {
+  sortOption: string;
+  updateSortOption: (sortOption: string) => void;
+}
