+++ src/types/Touched.ts
@@ -1,0 +1,4 @@
+export default interface Touched {
+  name?: boolean;
+  price?: boolean;
+}
