+++ src/types/WelcomeProps.ts
@@ -1,0 +1,3 @@
+export default interface WelcomeProps {
+  isNewUser: boolean;
+}
